+++ backend/src/main/java/com/ptit/qlphonghoc/facility/dto/FacilityOpeningActionRequest.java
@@ -4,7 +4,7 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
 public record FacilityOpeningActionRequest(
-        @NotNull @Positive Integer semesterId,
+        @NotNull @Positive Long semesterId,
         @NotNull String sourceType,
         @NotNull @Positive Long sourceId,
         @NotNull @Positive Long classroomId
